prepping for timestamps
Declaring it as a constant, and using nsdate unix epoch time. It's probably good practice to push timeSent to firebase when sendMessage function over in the viewcontroller is called. There's a lot you can do with that, and makes it much cleaner. Later you can convert epoch time to a user-friendly format and have the date and time sent displayed when user swipes over the messages (or you can just have them sit under/above the messages but that UI style is kind of cluttery.