Extension to FriendlyChat for PCI
Adds functionality in the completed project to write messages to the index,
and log user actions on these messages with the Firebase App Indexing API.

No changes in starter project, as the codelab will walk them through each
component that needs to be added.

I went ahead and also changed all the dependencies since all firebase
libraries will need to be 10.0.0 for these changes.

Change-Id: I2e1bfe262a10b2299bad1b339884d4ea43883ac5